split pinecone-client.ts with 3 sub files
# Conflicts:
#	src/pinecone-client.ts
